refactor(app): extract the connect wiring
The background health poll, the connect handler, cancelling an in-flight
connect and grid column drag-resize move to `wire::connect`. The tab
save/restore closures stay in `main`, so the cluster is cut around them.
