+++ app/src/wire/mod.rs
@@ -6,6 +6,7 @@
 
 pub(crate) mod browse;
 pub(crate) mod conn_form;
+pub(crate) mod connect;
 pub(crate) mod edit;
 pub(crate) mod find;
 pub(crate) mod grid;
